+++ common.h
@@ -5,7 +5,7 @@
 #include "aux.h"
 
 #define APPNAME "pingpath"
-#define VERSION "0.1.10"
+#define VERSION "0.1.11"
 
 #define MAXTTL 30
 #define COUNT  50