+++ styles.h
@@ -6,7 +6,6 @@
 #define PAD2 "16px"
 #define CSS_ID_WIN      "win"
 #define CSS_ID_APPBAR   "appbar"
-#define CSS_ID_MENU     "menu"
 #define CSS_ID_DATETIME "datetime"
 
 void init_styles(void);